third_party/blink/web_tests/external/wpt/html/semantics/forms/the-input-element/checkable-active-onblur-with-click-expected.txt

This is a testharness.js-based test.
[FAIL] Checkboxes should clear :active when the user tabs away from them while pressing it with a pointing device
  assert_equals: Checkboxes should not be :active after tab is used to change focus. expected null but got Element node <html><head>\n<meta charset="utf-8">\n<script src="/resourc...
[FAIL] Radios should clear :active when the user tabs away from them while pressing it with a pointing device
  assert_equals: Radios should not be :active after tab is used to change focus. expected null but got Element node <html><head>\n<meta charset="utf-8">\n<script src="/resourc...
[FAIL] Checkboxes should clear :active when the user tabs away from them while pressing the parent label with a pointing device
  assert_equals: Checkboxes should not be :active after tab is used to change focus. expected null but got Element node <html><head>\n<meta charset="utf-8">\n<script src="/resourc...
[FAIL] Radios should clear :active when the user tabs away from them while pressing the parent label with a pointing device
  assert_equals: Radios should not be :active after tab is used to change focus. expected null but got Element node <html><head>\n<meta charset="utf-8">\n<script src="/resourc...
[FAIL] Checkboxes should clear :active when the user tabs away from them while pressing the following label with a pointing device
  assert_equals: Checkboxes should not be :active after tab is used to change focus. expected null but got Element node <html><head>\n<meta charset="utf-8">\n<script src="/resourc...
[FAIL] Radios should clear :active when the user tabs away from them while pressing the following label with a pointing device
  assert_equals: Radios should not be :active after tab is used to change focus. expected null but got Element node <html><head>\n<meta charset="utf-8">\n<script src="/resourc...
Harness: the test ran to completion.
